Modifications to pragma_simd to allow arguments. Consistent work in ex3/async.cc. Added Python work to ex3/async.py

+++ madthreading/threading/thread_manager.cc
@@ -76,7 +76,7 @@ void thread_manager::check_instance(thread_manager* local_instance)
 
 //============================================================================//
 
-thread_manager* thread_manager::get_thread_manager(const uint32_t& nthreads,
+thread_manager* thread_manager::get_thread_manager(const int64_t& nthreads,
                                                    const int& verbose)
 {
     mad::thread_manager* tm = fgInstance;
